+++ docs/changelog/CHANGELOG.md
@@ -1,3 +1,27 @@
+# [2.33.0](https://github.com/Tallec7/neopro/compare/v2.32.3...v2.33.0) (2026-01-18)
+
+### Features
+
+- **remote:** add cloud-based remote control for mesh WiFi environments ([#379](https://github.com/Tallec7/neopro/issues/379)) ([159a250](https://github.com/Tallec7/neopro/commit/159a250ea8e6a6f7873f8ab44bffb2cf5cd1565e))
+
+## [2.32.3](https://github.com/Tallec7/neopro/compare/v2.32.2...v2.32.3) (2026-01-18)
+
+### Bug Fixes
+
+- **admin:** fix mDNS avahi service and admin demo mode detection ([#378](https://github.com/Tallec7/neopro/issues/378)) ([5cabd65](https://github.com/Tallec7/neopro/commit/5cabd65d7effc881f62ff35c5c0554e142f40cda))
+
+## [2.32.2](https://github.com/Tallec7/neopro/compare/v2.32.1...v2.32.2) (2026-01-18)
+
+### Bug Fixes
+
+- **central-server:** ESLint, tests, schema sync + hotspot wlan1 fix documentation ([#377](https://github.com/Tallec7/neopro/issues/377)) ([9ab1df7](https://github.com/Tallec7/neopro/commit/9ab1df7e030775171eea3ee9d8f1ccab2cddfcaa))
+
+## [2.32.1](https://github.com/Tallec7/neopro/compare/v2.32.0...v2.32.1) (2026-01-18)
+
+### Bug Fixes
+
+- **admin:** fix mDNS avahi service and admin demo mode detection ([#376](https://github.com/Tallec7/neopro/issues/376)) ([4d95bdc](https://github.com/Tallec7/neopro/commit/4d95bdc7415007e9c5948a38ea6055bb653af8ca))
+
 # [2.32.0](https://github.com/Tallec7/neopro/compare/v2.31.0...v2.32.0) (2026-01-18)
 
 ### Features